% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adult_mouse_brain_ffpe}
\alias{adult_mouse_brain_ffpe}
\title{Deconvolved cell-type proportions from STdeconvolve for a Visium data from the adult mouse brain along with the positions of their respective spots}
\format{
A list of 2 objects, each of 2264 rows and 2 columns: prop, which contains the 12 deconvolved cell-types proportions for each 2264 spots/pixels in a spatial transcriptomics experiment performed on adult mouse brain tissue tissue and pos, which contains the x and y-coordinates for each 2264 spots/pixels on an adult mouse brain tissue slide.
}
\source{
\url{https://www.10xgenomics.com/datasets/adult-mouse-brain-ffpe-1-standard-1-3-0}
}
\usage{
data(adult_mouse_brain_ffpe)
}
\description{
Deconvolved cell-type proportions from STdeconvolve for a Visium data from the adult mouse brain along with the positions of their respective spots
}
\keyword{datasets}
