% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deconProp}
\alias{deconProp}
\title{Deconvolved cell-type proportion matrix of spatial transcriptomic data from mouse olfactory bulb}
\format{
A data frame with 260 rows and 8 columns, that contains the eight deconvolved cell-types proportions for each 260 spots/pixels in a spatial transcriptomics experiment performed on mouse olfactory bulb tissue.
}
\source{
\url{https://science.sciencemag.org/content/353/6294/78}
}
\usage{
data(deconProp)
}
\description{
Deconvolved cell-type proportion matrix of spatial transcriptomic data from mouse olfactory bulb
}
\keyword{datasets}
