% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mOB}
\alias{mOB}
\title{Deconvolved cell-type proportions from STdeconvolve for a Spatial Transcriptomic dataset of the mouse olfactory bulb along with the positions of their respective spots}
\format{
A list of 2 objects, each of 260 rows and 2 columns: data, which contains the 8 deconvolved cell-types proportions for each 260 spots/pixels in a spatial transcriptomics experiment performed on mouse olfactory bulb tissue and xy, which contains the x and y-coordinates for each 260 spots/pixels on a mouse olfactory bulb tissue slide.
}
\source{
\url{https://www.science.org/doi/10.1126/science.aaf2403}
}
\usage{
data(mOB)
}
\description{
Deconvolved cell-type proportions from STdeconvolve for a Spatial Transcriptomic dataset of the mouse olfactory bulb along with the positions of their respective spots
}
\keyword{datasets}
