% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pos}
\alias{pos}
\title{Positions for spots from mouse olfactory bulb tissue}
\format{
A data frame with 260 rows and 2 columns, that contains the x and y-coordinates for each 260 spots/pixels on a mouse olfactory bulb tissue slide.
}
\source{
\url{https://science.sciencemag.org/content/353/6294/78}
}
\usage{
data(pos)
}
\description{
Positions for spots from mouse olfactory bulb tissue
}
\keyword{datasets}
